requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
google-auth
google-api-python-client
